refactor: コンポーネント化
