+++ React/additional_assignment2/src/types/todo.ts
@@ -1,0 +1,4 @@
+export type TodoType = {
+  id: number,
+  title: string
+}
